requirements.txt
# LLM Classification Fine-tuning Project Requirements (v2.0)
# Author: Morris Chou
# Course: EEP 596D DL (Fall 2025)
# Competition: https://www.kaggle.com/competitions/llm-classification-finetuning

# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Deep Learning - PyTorch
torch>=2.0.0
torchvision>=0.15.0

# Transformers and Fine-tuning
transformers>=4.35.0  # 需要較新版本支援 DeBERTa-v3
accelerate>=0.24.0
datasets>=2.14.0
peft>=0.6.0  # Parameter-Efficient Fine-Tuning (LoRA, etc.)
bitsandbytes>=0.41.0  # For 8-bit optimization
sentencepiece>=0.1.99  # For some tokenizers

# Training utilities
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: for experiment tracking
tqdm>=4.65.0

# Kaggle API
kaggle>=1.5.12

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
jupyter>=1.0.0
notebook>=6.5.0

# Evaluation
scipy>=1.9.0

# Optional: For notebook environments
ipywidgets>=8.0.0
